Add automatic detection of Anki path based on OS.
I have not tested this. On Linux it ought to work, but I can't verify it
since I have Anki installed via git and it is not on my path. On Windows
I'm going to try it in a moment. I don't have a Mac, so I cannot confirm
that it works.