bugfix: pathnames containing spaces do not work with drag-and-drop
The somewhat odd issue is that most terminal emulators (Windows and
Linux, at least) automatically add quotation marks around filenames
containing spaces, but do not add them if they don't. If they have
spaces, this results in a string containing the quoted text, which is
not necessary and throws an error.

Fixed by stripping single/double quotes from the beginning/end of input
file strings. I'm pretty sure this does not exclude any but the very
strangest valid filenames, but note that the handling does preclude the
use of even escaped ones at the end, as the trimmer will hit the
quotation mark before the escape.