fix bug where the EOS marker would show up before the EOT marker
Fixes GitHub issue #7.

Stripping consecutive/leading/trailing newlines before trying to loop
through the lines and add the markers avoids all sorts of possible
fiddly little errors and is *much* easier to understand. Maybe
theoretically a little bit slower, but if you're importing poems that
are long enough to make this an issue on a modern computer, there is
something seriously wrong with you. ;-)