refactor: use classes for poem lines
This will make it much easier to deal with grouping and multi-recitation.
My skills with index manipulation and range() and len() were already
stretched to the limit by the previous functionality!

For now, process_text() will continue to preprocess everything for us
without using any classes. If this needs additional work, it could be
refactored into the classes as well, particularly the bits about
indentation and end-of-line/stanza markers, but that seems like more
trouble than it's worth at the moment.